requirements.txt
outline-vpn-api
python-decouple
SQLAlchemy
loguru
PyYAML
aiogram